requirements.txt
Django==5.2.1
djangorestframework
markdown
django-filter
google-genai
django-extensions
python-decouple
python-dotenv